Default theme: removed logout applet
Here the logout applet is not very useful if another panel is already used
because panels of gnome, kde, unity, xfce, lxde, etc. already have a logout
applet. If no other panel is available, I guess the Cairo-Dock session is
used or the user knows how to enable this applet.

And this applet causes a bug with Unity's logout applet (LP: #1242112)