Implement positioning on Wayland for subdocks, menus and dialogs.
Separate approach is needed since it is not possible to set global window positions.

Fix positioning
===============

This includes the following implementation changes:

1. New wrapper around gdk_window_move_to_rect(), that can be called for setting up dialog and subdocks position.

2. Added the new function gldi_container_calculate_rect() that is now used by subdocks,
menus and dialogs to calculate the parameters to position themselves with the new method.

3. Refactor calculating aimed point for menus and dialogs. Use a common implementation provided by
gldi_container_calculate_aimed_point(). This is not used by subdocks: pointing arrows for them work
"magically" (I don't understand why, I assumed they would point to wrong locations).

4. Positioning subdocks, menus and dialogs now uses the above functionality that works both on X and Wayland. Arrow does not work for vertical
docks, but I'm not sure if it works on the master branch as well.

Does not work: moving an applet to a different dock (subdock positioning is
messed up after and needs Cairo-Dock to be restarted); I'm guessing this
should be fixable by setting the parent to the new dock at the right location.

Additional problem: this would normally need GDK 3.24 at least. I've added
a hack (based on code in gtk-layer-shell) to make it work on GDK 3.22, by using
an unexposed internal function, since I did not want to upgrade my GDK
libraries (this is in the previous commit).
Probably this should be removed in the future.

Separate positioning logic between X11 and Wayland
==================================================

New code for positioning menus, subdocks and dialogs should work on both Wayland and X11.
Now, by default, it is only used on Wayland, and the old code is used on X11 to
avoid regressions. There is a new cmake option (-Duse-new-positioning-on-x11) that
can be used to test the new code on X11 as well; I've been using Cairo-Dock on X11
in the past few months with this option without any problems. At a later point,
this options could be made the default to simplify things.

In the future, things might be handled differently, either
 - ideally, the same positioning code could be used in both cases (using the container_calculate_rect() and gldi_container_move_to_rect() functions)
 - alternatively, backend-specific positioning could go in the backend managers

Additional small changes:
=========================

 - Do not check for dock visibility on Wayland, it does not work. There is no suported way to get window positions.
 - Add explicit command line option to choose between X11 and Wayland. This does not work properly with OpenGL enabled, EGL tries to use X11 by default; #10 should fix that.
 - Add command line option to disable layer-shell.
 - Improve closing dialogs. Handle when the compositor closes a dialog on Wayland, fix an issue with clicking on the dialog:
this made the main dock "freeze", by stopping it receiving any GDK events. It seems that the dialog receiving a mouse button
press without a corresponding release event confused GDK / GTK. This happened if closing the dialog as a response to a mouse
 click. Now, we wait until the mouse button release event to close the dialog.