hacks to provide a function necessary for subdock placement on GDK 3.22
The gdk_window_move_to_rect() was added in GDK 3.22, but is only part of the
public interface since GDK 3.24. Since I only have 3.22, this is necessary to
be able to test and develop. This probably should be removed in a future
release and instead a strict dependency on GDK 3.24 could be required.