Parse app ID under Wayland as well
Moved the functionality for this from cairo-dock-X-utilities.c to
cairo-dock-windows-manager.c (as a new function: gldi_window_parse_class()),
this is now used by both the X and Wayland backends. This fixes issues by
not recognizing some apps (e.g. Firefox and gnome-terminal if using XWayland)
due to the app ID not matching the expected case.

Also re-wrote the logic for identifying and deciding when an icon should be
shown for a foreign-toplevel. Signals are now always emitted by the done event
handler.