Impementation to manage windows on Wayland
Using wlr_foreign_toplevel_management_unstable_v1, available on wlroots-based compositors.

See https://github.com/swaywm/wlr-protocols/blob/master/unstable/wlr-foreign-toplevel-management-unstable-v1.xml

Basic actions work:

 - adding / removing icons when windows open / close
 - matching to launchers
 - activate, minimize, close actions
 - set icon position for minimize animation
 - filtering out dialogs (they should not appear separately)
 - selecting a window interactively to grab the app id (pick_window())

Not working:

 - changing workspaces / figuring out which workspace a window is on
 - getting the position / size of windows (not supported in this protocol)
 - subdocks show up as icons, unless layer-shell is used

Tested on Wayfire https://github.com/WayfireWM/wayfire/

Notes:

 - Since we do not know the real position of windows, we "put" them in the middle
of the screen for now; this way they will not be considered overlapping with the
docks.
 - Moved the functionality for parsing app IDs from cairo-dock-X-utilities.c to
cairo-dock-windows-manager.c (as a new function: gldi_window_parse_class()),
this is now used by both the X and Wayland backends. This fixes issues by
not recognizing some apps (e.g. Firefox and gnome-terminal if using XWayland)
due to the app ID not matching the expected case.
 - Functionality to set minimize position / thumbnail area now requires the parent
dock as a paramter since this location is set in relative positions; the API is
changed accordingly to this.
 - The gldi_window_pick() functionality on Wayland is implemented by getting
notified of the window that is activated by the user. Also, a separate dialog is
created since it is not possible to globally change the pointer to a crosshair, so
the dialog is used to give some feedback to the user. This requires passing the
parent window as a parameter, changing the API in this case as well.