wayfire-integration: use a proper JSON implementation
This makes the implementation more robust, but adds a build dependency on nlohmann_json. However, this is already a dependency of Wayfire, so this is likely not a problem (and does not add any runtime dependency). Note that this also means that parts of the code are compiled as C++.