user-icon-manager: fix loading of custom subdocks
Previously, when loading a launcher or separator, if it was part of a nonexistent container, it was created as a new root dock. If later encountered as a subdock, the type was changed to a subdock. This does not work with the new positioning and Wayland, since subdocks need to be created as popup windows with a correct parent set. This is not easy to achieve later, so instead, ensure that all subdocks are loaded first, before loading launchers and separators.