remove .exe and .py suffices from app IDs
These are already removed in some, but not all cases, and can be potentially problematic when matching icons or desktop files.

Specific example: winecfg.exe with no .desktop file and winecfg.png as icon.

This can cause issues if two apps use IDs that only differ in their suffix, but this is unlikely.