class-manager: handle apps with non-unique command or StartupWMClass
This can happen since the StartupWMClass can have any value (e.g. if an app has multiple launchers for some reason). Also the command in Exec= can be non-unique if the same app has multiple launchers or uses an interpreter that is not parsed by us (e.g. python).

To handle this, we only assume that the app-id (as reported by the WM) and the .desktop file ID are unique (if multiple apps use the same app-id, we cannot do much to distinguish them, although there is already some special casing under X11 where a separate "name" might also be provided).