class-manager: avoid creating a new class accidentally
cairo_dock_get_class () will always create a new CairoDockClassAppli if it cannot find the class given as an argument, however this seems not to be desired in a couple of places where call to this function was already followed by a NULL check. Change it to _cairo_dock_lookup_class_appli () instead which will return NULL if the class is not found.