class-manager: rework API
Make cairo_dock_register_class_full () private and instead offer a new version, cairo_dock_register_class2 () that is hopefully less confusing.

Also fix a bug with not refing GAppInfo objects and a possible bug when apps are not found based on a custom WMClass.

Fix a minor bug when an icon's label is incorrectly updated when setting it in a launcher's config.