user-interaction: treat all parameters as URIs for DnD targets
g_file_new_for_path () will just prepent the CWD (of Cairo-Dock!) to any string it does not understand, which is not very useful, we anyway expect the drag source to provide absolute paths. So it is better to exclusively use g_app_info_launch_uris () which will pass parameters unaltered in most cases.