utils: rework launch_command() and friends
Create versions that take an argument vector. Since the g_spawn* functions use this internally, this way we can avoid a "roundtrip" with g_strdup_printf (...) and g_shell_parse_argv (...) and any issues with quoting and escaping.

Rework the async version to use g_spawn_async() instead of launching a thread and calling system(). This makes it cleaner to set the working directory as well.

Added an option to the async version for launching GUI apps: this uses GdkAppLaunchContext internally to set DESKTOP_STARTUP_ID / XDG_ACTIVATION_TOKEN.