surface-factory: ensure that files are opened with O_CLOEXEC
This is a bit paranoia, since in most cases we load images and launch apps from the same (main) thread, but in theory, these can happen in a GldiTask or similar, so better to be sure. Interestingly, cairo version < 1.18.2 does not ensure this, while gdk-pixbuf does not do so at all.