class-manager: process desktop actions separately
This was removed before, but is necessary to control launching these in more detail as there is not support for them in g_desktop_app_info_launch_uris_as_manager ().

Create a new object that encapsulates GDesktopAppInfo and use it to store it for a class and with the icons.