build: better detection of systemd for default value
The "enable-systemd-servie" option controls whether a systemd unit is generated and installed. If this is not set, we try to auto-detect if the host is running systemd and decide to install our unit based on this.

Before, we used pkg-config for this, but this is not necessarily a good idea, since it depends on having the systemd dev packages installed (and thus broke autodetection on e.g. Ubuntu plucky). Instead, we can run "systemd-notify --booted" which reports whether systemd is running (and if the command is not found, we can know systemd is not present anyway).