application-facility: improve calculating minimize positions
In the case when an icon is not already in the dock (using the setting "show only minimized"), we should use the same algorithm to calculate minimize positions as when actually minimizing. This way, the position will be consistent with the minimize animation of the WM.

Also, we should use the real icon size for the minimize area to match better where the icon will appear.