dialogs and menus: set a background if no renderer is available
This improves the legibility of any text displayed. The background will be a simple rectangle that covers the whole dialog or menu window.

Note that renderers will only be unavailable if we could not load the dialog-rendering plugin -- in this case, we will very likely display a warning dialog and having a transparent background can make it very hard to read it.