dock-manager: do not show subdocks when they are created
Previously, when a subdock was created, it was immediately shown and then hidden. This was convenient to ensure that it's GdkWindow was realized so we could access it when needed. However, this will lead to a crash or freeze on Wayland if the subdock's parent is not shown (its surface is not mapped) -- this can happen for nested subdocks.