wayland-manager: do not commit when updating surface input shape
This can cause problems if we commit at a point where the surface state is not ready. E.g. having updated the buffer scale, but not have provided an appropriate new buffer yet. Also, this should not be necessary, any time we need to update the input shape, we are also redrawing the dock which will commit the surface.