glx: implement making current an "offscreen" context here as well
We use the static context created during the initialization. However, we still bind the surface of the main container, as making a context current without a surface is only supported on later OpenGL versions (and it is a bit unclear when exactly). The caller should avoid rendering without binding an FBO / texture.