class-manager: reword GldiAppInfo, bring back command line processing
This allows us to use the new custom launch facilities instead of relying on GAppInfo functions. This is necessary to be able to launch apps as transient services on systemd as opposed to putting them in a transient scope after they launch.

This also makes GldiAppInfo opaque, as other parts of the code do not need to access its internals.