icon-factory: add fYAtRest parameter for icons
This is currently used to set minimize position for icons, so that the minimize animation can start exactly from the top of an icon (at their "at rest" position). This is needed as icon vertical positions can be adjusted by the renderers in a non-trivial way, so it is complicated to calculate this when setting minimize positions.