dock-factory: maintain a separate list of applets attached to a dock
Some applets might detach their icons (e.g. Status-Notifier, Cairo-Penguin), resulting in a main dock being destroyed (if no other icons are left) and then a crash when they reference the dock again. We can prevent this by explicitly noting which applets are attached to a dock and preventing its destruction as long as this is not an empty list.