dock-facility: handle the mouse leaving (sub)docks with complex shape
For some dock renderers, the input shape we set does not match their real shape, so we manually check whether the mouse is outside. However, code to trigger hiding the subdock in these cases has not been working for some time. This adds the necessarily changes to re-enable this functionality.

Also add more comments about leave events in dock-factory, since the situation is quite complex.