logout-backend: add DE-independent functions for shutdown, logout, etc.
These were previously part of the "logout" applet, but since other components might want to use them, it is better to have them here (and it's not worth having a separate plugin for them). DE-specific functions still take precedence.

Note: the code is taken from the "logout" applet, with adaptations to use GDBus and to also use DBus methods for logging out and locking the screen.