desktop-manager: add purely systemd-based session for newer gnome-session versions
This means relying on systemd to start all required components (including Cairo-Dock and the WM) and thus completely removing the RequiredComponents= setting from the session files.

Note: we also add systemd units for Compiz and Metacity as these might not be provided by their respective packages, but we need them in this case.