desklet-factory: interpret negative position as right / bottom alignment
Giving a negative coordinate in the settings was interpreted as distance from the right / bottom screen edge, but still applied to the top left coordinate of a desklet. This does not work well for desklets that resize themselves, e.g. the systray applet. Ensure instead that the desklet's container is aligned to the bottom / right as well.