dock-manager: fix double free and memory leak
Calling gldi_container_update_polling_screen_edge () will typically trigger the backend to query all root docks in s_pRootDockList. When freeing all docks at once, this can be a problem, as this list is not updated anymore. We can avoid this by freeing the list before the dock table.

Also free a possible memory leak with not freeing pVisibilityData.