launchers: only match StartupWMClass for custom commands
If a launcher is not linked to a .desktop file on the system, do not attempt to do so if a custom window class is also given. This allows defining launchers for the same executable but with different window class (for apps that allow setting the window class as a command line argument, e.g. terminals).