module-manager: keep track of active modules based on the config
Instead of updating the config file from the modules that are active, be more careful and just add / remove elements to the list as modules are activated or deactivated. This means that missing modules will not be removed. This fixes the case when if e.g. running Cairo-Dock without OpenGL, plug-ins that require it would be disabled if any other plugin was added or removed.