module-instance-manager: process config file upgrades earlier
Before, updates to a plug-in's config file (i.e. when a plugin has a new version) were only updated after attempting to read when loading an instance. This could result in some values set inconsistently, as the CD_CONFIG_GET_* functions apply their own default values which can be different than the values set for newly added keys in the updated template. Avoid this situation by upgarding the config file before attempting to read it.